requirements.txt
streamlit>=1.32.0
numpy>=1.26.0
scipy>=1.12.0
plotly>=5.20.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
